Support the case-insensitive symbol names.
This commit is to support that the symbol names that are case-insensitive from
the `/usr/include/X11/keysymdef.h` - `XK_something` are shown as occupied (red
color) on the page by the `i3keys web`. Because sway supports the symbol names.

The symbol names are such as `Semicolon` for the `XK_semicolon` in the
`keysymdef.h`.